Fix #139248: Update IME Data with GHOST_kEventImeComposition Event
Before e4aa758d70 we kept a pointer to ghost IME data, but with that
change we instead maintain a local copy of that data instead. This
means that we have a copy of it with GHOST_kEventImeCompositionStart,
but we don't update our copy when we get GHOST_kEventImeComposition.
This is where (at least on Windows) when we get the composition result,
the final string to be inserted.