Fix: Cursor options display inactive in Sculpt Mode
Prior to this commit, the cursor overlay options were tied to the
`has_overlay` setting for brushes. This check used the `mtex` property
to examine the mapping mode of the associated texture.

This should not apply to the overall cursor overlay settings, as they
are not dependent upon a texture being loaded. Additionally, Sculpt mode
uses the `mask_mtex` property instead of the `mtex` property for the
pimary texture information.