Python API: Make bpy.app.timers.register() thread-safe
Currently, if you want to schedule a function to run on the main thread,
you're forced to use a timer to poll a background queue for new tasks.
As a result, you're forced to choose a "good" compromise between having
your tasks executed in a timely manner on the main thread vs slowing
down the main thread by having an interval that's too low.

Moreover, enqueuing work from a background thread to the main thread is
a common operation (common enough that the docs suggest to use a timer
to accomplish this task), so every extension needs to manage their own
queues and their own polling code to accomplish this same task.

This commit makes bpy.app.timers.register() thread-safe, so that it can
be directly called on a background thread to enqueue work on the main
thread.

To keep the implementation simple, the core timer datastructure remains
unchanged. Instead, registration adds entries to a thread-safe queue,
which is drained into the usual GlobalTimer structure on the main
thread.