Fix #134153: Do Not Snap To Two Columns for Tools With Unaligned Buttons
The current behavior for Toolbars is to snap to nothing, then one
column of icons, then two columns of icons, then a single column of
icons with text. However the tools can only be shown in two columns
if any of them are connected. Two that are connected vertically can
be shown in two columns, and four in a group can be shown as two by
two. In the Nodes editors there are (usually) just three items and
each are separate. This means that they remain in one column at the
two-column snap position. This makes it hard to find the line to resize
the region. This PR removes the two-column snapping position when not
applicable.