EEVEE: Add deferred texture loading
This adds a new function to query GPUtexture from an
Image datablock without actually creating them.

This allows to keep track of all the texture that
needs to be loaded and defer their loading in
end_sync. The texture are then only used in the
next sync. This is because we do not want to stage
the texture for drawing as it would require a
valid texture.

Having a vector of all the textures to be loaded
will eventually be helpful in making the
texture loading multithreaded. Currently, it is
a bit difficult given the need of a valid GPUContext
per thread.

TODO:
- [ ] Bypass deffered loading on animated textures