UI: Remove Material Slot Button is properly disabled in edit mode
Disables the remove material slot button in edit mode properly. Previously the remove material slot button still appeared to be active while in edit mode. Attempting to use the button would result in an error "Unable to remove material slot in edit mode."

In the code a check was added for edit mode during operator execution but never added to the polling function. This commit adds that check.