Use Color Temperature prop subtype for light temperature
The light objects' Color Temperature prop introduced in a12bce039f
currently is a temperature subtype. This means it is expressed in °C
if the user set that as scene unit, but a color temperature should
always be expressed in Kelvin.

This commit changes the subtype to Color Temperature to get this
behavior.