Fix #139705: Generate undo step after cancelling node.join_named
In macro operator `NODE_OT_join_named`, panel is called to name the
frame after execution of `node.join`. This panel returns
`OPERATOR_INTERFACE` to `wm_macro_invoke_internal/wm_operator_invoke`.
This prevents from generating an undo step for node.join because retval
is not `OPERATOR_FINISHED`. Hence, `wm_operator_finished` is not
called which actually calls the function to create undo step. To fix
this, adjust condition in `wm_macro_end` so that retval is "FINISHED"
when first operation is operation is executed successfully while later
is in `OPERATOR_INTERFACE` state/