Fix #138175: EEVEE crash with multiple materials in sculpt mode multires
`BKE_id_material_used_with_fallback_eval` doesn't account for the case
the evaluated mesh is empty and `SubdivCCG` is used for drawing instead,
which only happens for workbench/EEVEE in sculpt mode. Instead of
further complicating the logic, change the multires modifier to produce
a copy of the input mesh with the SubdicCCG attached, rather than that
mesh being empty. This makes things like detecting the max material
index work "for free."