EEVEE: Use simpler hash for ObjectKey
Use the C++ "get_default_hash" instead of the C GHash functions. These
are inlined and also simpler, which is typically good for our open-
addessing hash tables. In a few scenes with many objects/instances,
I observed a 2% improvement in playback FPS.

Also the less than operator was unused. I figured this was probably
used in a `std::map` or another container that required the operator
at some point.