Cycles: Introduce Optimization Information Status
With these changes, you can now mark in Blender Cycles
devices which are expected to work as performantly as possible,
and devices which were not optimized for some reason.
For example, because the device was released after the Blender
release, making it obviously impossible for Blender developers to
optimize for devices in already released unchangeable code.
This is primarily relevant for the LTS versions, which are supported
for two years and require proper communication about optimization
status for the new devices released during this time.

To keep compatibility with the old behavior, all devices are
marked as optimized for performance by default.
No functional changes.