Cycles: oneAPI: Mark devices by their optimization status
Currently, several oneAPI devices can be accessed with the
default Blender configuration and even more with disabled
device filtering on the Blender side. Some of these devices
are optimized as much as possible, while others are not
optimized, especially if these devices were released after
the latest or LTS Blender version.