Fix #133533: Invert scroll direction in render pass menu
When ctrl+scrolling down in the render pass menu of the image editor, the
menu would select the pass above the current one while supposed to
select the one below it. Ctrl+scrolling up would also scroll in the
wrong direction.

This has been fixed by inverting the scroll direction.