VSE: Add preference for matching scene framerate to movie
With #135494, automatic changes to scene framerate are more visible,
but there should still be a way to control whether/how it is triggered.

This patch adds a user preference to control such behavior when
adding an initial strip to the timeline, using a new operator,
"Match Scene Frame Rate to Strip" (`match_scene_to_strip`).

Options are:
- Prompt (default)
  - A popup box will appear if the scene framerate does not match the
    imported movie. The user can then choose whether or not to go
    through with the change.
- Always
  - The scene's framerate will always be updated to match the scene. An
    info report in the status bar mentions when this occurs.
- Never
  - The user must manually change FPS through output properties. No
    changes or warnings ever occur.

This has the added benefit of preventing automatic frame changes for
Python scripts and giving them more control to do so on their own terms
with the new operator.

The patch does not touch automatic view transform changes, since they're
already a bit of a hack, and would be better dealt with by supporting
those view transforms natively.