Fix #139846: Plane brushes behave incorrectly with constant falloff
When constant falloff is used, BKE_brush_calc_curve_factors doesn't do
any extra filtering of the given distances. The Plane brush previously
didn't filter the distances, leading to incorrect deformations when the
constant falloff was used.

To fix this, add an explicit filtering step for the plane brush.

---

Discussion:

Currently unsure about if we want to change the behavior of
BKE_brush_calc_curve_factors in 4.5 to avoid this filtering and simply
assert on this condition. It seems like the more correct thing to do
overall but arguably skirts the line of what is and isn't a fix and may
be risky for inclusion in the stabilizing branch.