Physics: Use smoke heat for calculation of Adaptive Domain size.
With previous implementation only Smoke density and fuel,
were taken during calculation size of Adaptive Domain.
As Heat grid is also used for displaying smoke in Blender with Fire and Smoke,
we should take heat into calculating of the domain size.

With such solution we are not loosing information about heat,
and as a result Adaptive Domain have the same simulation results
as Fixed Size Domain.

Fixes: #107679