Fix #139740: expose Bone Wireframe Opacity in all shading modes
This option is not only in effect in Wireframe shading mode, but also if
the armature object `Display As` is set to `Wire`.
Having to go to Wireframe shading mode only to set this option was
confusing, therefor expose it in all shading modes.

NOTE: since Overlay Next, we seemingly fixed the issue of custom bone
shapes **not** taking this into account which made it seem like bones
using them where invisible in 4.4 (where they were visible in 4.3) --
with no obvious way to get them back