Fix assert in bmo_dissolve_edges_exec()
This restores the BM_vert_is_edge_pair(v) check that was present prior to !134017.
In that PR, the edge pair check was moved up, to a previous iteration pass, that checked the angle thresholds.
However, even though pairs were checked before, the process of performing edge merges might change a neighboring vert such that it is no longer an edge pair.  Therefore it is necessary to check a second time.