Fix #140001: EEVEE: Missing refresh when loading texture
This happens because the viewport refresh happens only once and
only if the frame contains a texture upload.

There are multiple fix to this issue:
- Call `inst_.sampling.reset` at the begining of a sync if
previous sync had texture loading.
- Call `inst_.sampling.reset` at the end of a sync if sync has
texture loading.

But given we can also have issue with shadows update, this
patch follows the update detection from shader compilation
and uses a timestamp.