Fix #139882: Vulkan/USD: Crash on AMD GPUs
When USD sends work to the GPU (dispatch) incorrect datatype conversions
can send incorrect vulkan commands. This happens on AMD GPUs. This PR
will patch USD to limit the max allowed limit. A better solution would
be to only use uint32_t in this function, but there is no vec3u data
type and due to not able to compile locally I didn't want to do to many
changes.

**Note**

- I wasn't able to compile OpenUSD locally so wasn't able to test the
  change
- Due to previous bullet I have not been able to validate that this is
  the only needed change.