Fix #139815: Vulkan: Hang when loading minimized windows
e93479216934dd23a87c9a34cc13de1be30182b7 introduced a workaround for
NVIDIA windows where NVIDIA drivers fail to allocate swapchain images
of 0,0 resolution. The fix was to clamp it to 1,1. Now AMD driver tells
Blender that the created swapchain is suboptimal, blender tries to
recreate the swapchain, but the resulting swapchain will always be
sub-optimal.

This PR limits the clamping to NVIDIA drivers only.