Mesh: Extend GPU batch limit from 32 to 64
Some refactors (!140156) require adding a few more batches,
but we're already at the limit for how many we can add. This
commit makes the batch flag into a 64 bit integer. It might be
worth refactoring this more to use some system besides a flag,
but for now this simple change is best.

The atomic combination of requests wasn't actually required
and isn't supported for 64 bit integers, so it's just removed here.