Fix #140162: Root and Sphere strength curves corrupt mesh
Introduced with 23951e1b124b0e5d6431dfa987e10fbdaec1ce74

When mapping to a root or sphere curve, the square root of the relative
ratio of the current distance to the overall brush is taken. This
results in setting values to NaN when the distance is greater than the
radius, as the square root of a negative number is NaN.

To fix this, skip any computations where the factor is already known to
be 0.

---

An alternative to this is to apply `std::abs` before calling `sqrtf`,
though this potentially makes the algorithm less clear. It seems like
what we ideally need for this and other functions is to skip processing
once the factor will have no effect, but that is a much bigger change
than something that should be included in this fix.