Compositor: Add Signed Distance Field node
This patch adds a new Signed Distance Field node. The node takes in a
mask and performs an euclidean distance transform to compute the
distance to the nearest mask boundary pixel. The distance is negative
inside the mask. The node also returns the integer coordinates of the
nearest mask boundary pixels.