Keymap: Introduce option for timeline navigation
The current shortcuts to navigate the timeline are inconsistent both
within Blender and with UI standards. Some operators will use up and
down to mean after and before, while others are reversed:
- Jump to Keyframe uses Up Arrow for next frame, Down Arrow for
  previous.
- Jump to Strip uses Page Up for previous frame, Page Up for
  previous (this was changed in e1c9704108).

The Up for Next, Down for Previous behavior has been in Blender
forever and many users will be very much used to it. On the other
hand, it is not intuitive and users of other software will prefer the
more common shortcuts.

This commit introduces a preference to keep the old behavior, while
reversing it by default, so that Up means Previous, Down means Next.