Online Assets: Let downloader block asset library loading
Adds initial bits to let the downloader communicate status information
to the asset library loading.

Adds 3 BPY functions to let the downloader communicate status
information to the asset system:
- `WindowManager.asset_library_status_is_loading()`
- `WindowManager.asset_library_status_finished_loading()`
- `WindowManager.asset_library_status_cancelled_loading()`

They all take the library URL, so we can store the status per URL, and
display per library status accordingly.

The first function also takes a timeout (defaulting to 0.3 seconds). If
none of the status report functions is called for this duration, the
download is considered failed (TBD: should we differentiate between
cancelled and failed?). For this to work at least the `_is_loading()`
function should be called repeatedly while the download is running.

All communication happens on the main thread.

---

The additions to the downloader are rather quick and dirty. We need to
figure out the best way to track and query status there still.

Currently we just let the asset loading wait until the downloader is
done. Next step is to display assets progressively as the downloader
finishes processing individual pages.