Fix #139870: Cycles: Some objects with normal maps leak light
This was broken by the refactor of the microfacet code to no longer check the
"geometric normal", which in reality was the smoothed normal.

Since the logic is now the same for all closure types, it seemed weird that the
light leak only affects Microfacet closures, not Diffuse.

Turns out that for diffuse closures, the relevant paths were rejected by the
initial hemisphere check in the smooth bump terminator code, which also
incorporates the smoothed but non-bump/normal-mapped normal sd->N.

So, we can detect and prevent the new light leaks by extending this check to
all closure types for the eval case. Sampling already has stricter checks, so
this doesn't apply there.

With this change, we can revert the two test cases back to their pre-refactor
version. In hindsight it was a mistake to just shrug off these changes as
okay, I should have looked closer into the difference.