Fix #140376: Modeling: Better message for showing mesh data transfer
Previously in 308292b5d73d5fa47ba757a83299350c1f328810 an error message
was added to notify that some selected objects are not valid for mesh
data transfer, that message didn't take account that src==dst is also a
case where the transfer is skipped, that doesn't mean the object is
invalid. So update the logic for this.