Anim: remove Compositor action selector from Scene properties
The Compositor node tree is no longer embedded in the Scene, so having
its action and slot selector in Scene properties doesn't properly
represent the data model anymore.  For example, the node tree could
in theory be shared across multiple scenes as a standard setup for a
project.

Note that #140347 should land before this, since it adds the action/slot
selector for node trees to the node tree editor's side-panel. Otherwise
this PR makes the compositor's action/slot selector completely
unaccessable.