Core: always use current open file for `blend_name` variable
Previously `blend_name` was using the same path to get its file name
as `//` used. However, the choice of path for `//` seems to be
bespokely chosen at each call site.  There may be a logic to it, but
it's not especially predictable for users.

This PR changes `blend_name` to instead always use the currently open
("global") file, making it well-defined and predictable for users.

This also prepares better for PR #139438 for 5.0, which adds
`blend_name_lib`, which always uses the blend file that the ID owning
the path is linked from. Over-all this should be much more predictable
and controllable for users.